requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
black==21.9b0
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
constantly==15.1.0
cryptography==36.0.0
daphne==3.0.2
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.17.1
ghp-import==2.0.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
incremental==21.3.0
Jinja2==3.0.1
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
Markdown==3.3.4
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.2
msgpack==1.0.3
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.9.0
Pillow==8.3.2
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.10.0
PyJWT==2.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.4.1
pyyaml_env_tag==0.1
regex==2021.9.24
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.1
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
watchdog==2.1.5
winrt==1.0.21033.1
zipp==3.5.0
zope.interface==5.4.0
